Norwegian translation by Jacob Frøkjær
Developed as part of the Google Code-In program in December 2011.
Translation author email address unknown. Communication only happened via
the GCI web page. I was a mentor on this task.

More details here:
http://www.google-melange.com/gci/task/view/google/gci2011/7130462

I've done some minor edits to this text. It's decent, but it could still use
some more editing to be considered a high quality translation.
Unfortunately the GCI program is finished for this time.